backend/requirements.txt
# Child Growth Monitor Backend
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1

# Database
# psycopg2-binary==2.9.9  # PostgreSQL adapter - commented out for dev
SQLAlchemy==2.0.36

# Data processing
pandas==2.2.3
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# API and validation
marshmallow==3.20.2
webargs==8.4.0

# Security (compatible versions)
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Task queue (simplified)
celery==5.3.4
redis==5.0.1

# Utilities
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
